+++ src/main/java/net/silvertide/pmmo_farmers_compat/mixin/CuttingBoardProcessStoredItemMixin.java
@@ -1,8 +1,8 @@
 package net.silvertide.pmmo_farmers_compat.mixin;
 
-import net.minecraft.core.BlockPos;
+import com.llamalad7.mixinextras.sugar.Local;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.level.Level;
 import net.silvertide.pmmo_farmers_compat.util.EventUtil;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -12,12 +12,9 @@ import vectorwing.farmersdelight.common.block.entity.CuttingBoardBlockEntity;
 @Mixin(CuttingBoardBlockEntity.class)
 public abstract class CuttingBoardProcessStoredItemMixin {
     @ModifyArg(method = "lambda$processStoredItemUsingTool$2(Lnet/minecraft/world/item/ItemStack;Lnet/minecraft/world/entity/player/Player;Lvectorwing/farmersdelight/common/crafting/CuttingBoardRecipe;)V", at = @At(value = "INVOKE", target = "vectorwing/farmersdelight/common/utility/ItemUtils.spawnItemEntity(Lnet/minecraft/world/level/Level;Lnet/minecraft/world/item/ItemStack;DDDDDD)V"), remap = false)
-    public ItemStack modifySpawnItemEntityArg(ItemStack resultStack){
-        Level level = ((CuttingBoardBlockEntity)(Object)this).getLevel();
-        if (level == null || level.isClientSide) return resultStack;
-        BlockPos pos = ((CuttingBoardBlockEntity)(Object)this).getBlockPos();
+    public ItemStack modifySpawnItemEntityArg(ItemStack resultStack, @Local(argsOnly = true) Player player){
         for(int i = 0; i < resultStack.getCount(); i++) {
-            EventUtil.postPlayerCraftEvent(level, pos, resultStack);
+            EventUtil.postPlayerCraftEvent(player, resultStack);
         }
         return resultStack;
     }